Fix rejection sampling
The original rejection sampling (sample_from_nde) was doing 20k
iteration long pilot run to get the maximal log-likelihood. Then it was
taking only 1 sample. This commit fixes that.

However, for applications on the limited experimental (neuroscience)
data, learning the proposal is not needed at all, as p(theta) is just
known (empirically, i.e. from the data... and it is often not gaussian
at all). So, fixed rejection sampling is commented out.